Initial implementation of <MUIPlacesAutocomplete> component
This commit is the initial implementation of the <MUIPlacesAutocomplete>
component. The component leverages the <Autosuggest> component from the
'react-autosuggest' package as well as the Google Places API Web
Service/Google Maps JavaScript API to provide autosuggest/completion
functionality when searching for a place. The main feature of the
<MUIPlacesAutocomplete> component is that it provides a Google Material
Design feel to the rendered elements via the Material-UI library.

Moving forward I look to iterate on the implementation by querying for
constructive criticisms as to how I can improve my JS/front-end skills
as well as how I can make the component more useful for other consumers.

Included is a demo which people can run locally to see the component in
action.
	modified:   .babelrc
	modified:   .eslintrc
	modified:   .gitignore
	modified:   demo/Demo.jsx
	modified:   demo/index.html
	modified:   package.json
	modified:   src/MUIPlacesAutocomplete.jsx
	new file:   src/images/google-logo-on-white-bg.png
	modified:   src/index.js
	new file:   test/api-key.js
	new file:   test/setup-dom.js
	new file:   test/test.jsx
	new file:   test/test.jsx.snap
	new file:   test/testHelper.js
	new file:   webpack.config.js
	modified:   yarn.lock