Initial implementation of <MUIPlacesAutocomplete> component (#2)
* Initial implementation of <MUIPlacesAutocomplete> component

This commit is the initial implementation of the <MUIPlacesAutocomplete>
component. The component leverages the <Autosuggest> component from the
'react-autosuggest' package as well as the Google Places API Web
Service/Google Maps JavaScript API to provide autosuggest/completion
functionality when searching for a place. The main feature of the
<MUIPlacesAutocomplete> component is that it provides a Google Material
Design feel to the rendered elements via the Material-UI library.

Moving forward I look to iterate on the implementation by querying for
constructive criticisms as to how I can improve my JS/front-end skills
as well as how I can make the component more useful for other consumers.

Included is a demo which people can run locally to see the component in
action.
	modified:   .babelrc
	modified:   .eslintrc
	modified:   .gitignore
	modified:   demo/Demo.jsx
	modified:   demo/index.html
	modified:   package.json
	modified:   src/MUIPlacesAutocomplete.jsx
	new file:   src/images/google-logo-on-white-bg.png
	modified:   src/index.js
	new file:   test/api-key.js
	new file:   test/setup-dom.js
	new file:   test/test.jsx
	new file:   test/test.jsx.snap
	new file:   test/testHelper.js
	new file:   webpack.config.js
	modified:   yarn.lock

* License project under MIT license

While the 'package.json' file already listed the license as the MIT
license this now officially licenses the project with the MIT license.
	new file:   LICENSE.md

* Allow Google API key to be provided as environment var for test

This commit allows the Google API key to be provided as an environment
variable for testing purposes. This will make it easier to perform
testing in a CI/CD pipeline.

This commit is in response to requested changes for PR #2
	modified:   test/api-key.js

* Add README.md

This commit adds the README.md file which includes content for
installing, setting up, using the component, and running a demo of the
component amongst other info.
	new file:   README.md

* Add `contributors` section to `package.json` - Add Matt as contributor

This commit updates the `package.json` file to add a `contributors`
section. We also include Matt Bailey as a contributor to the project.
	modified:   package.json