Update docs after renaming repo/package to mui-places-autocomplete
Having a repo/package/module named
google-places-material-ui-autocomplete seemed a little tedious to write
out each time you needed to so I renamed the repo/package/module to
mui-places-autocomplete.

I've also added keywords (all the keywords!) to the 'package.json'. I
might have been to excessive. Not sure if this is good or not. Hopefully
npm/yarn will "find us" frequently based on the keywords. If not we
ought to optimize them (reduce the number?).
	modified:   README.md
	modified:   package.json