Point package entry point to 'dist/index.js'
While testing if the package is ready to be published it was found that
the entry point for the package wasn't sufficient. It was pointing to
'src/index.js' which caused other projects to have compilation errors as
some of the syntax was unknown (as it was ES6 syntax). By changing the
package entry point to 'dist/index.js' which contains the transpiled
source the module should be ready for immediate usage after installing
the package.
	modified:   package.json