Make <MUIPlacesAutocomplete> safe to render on the server (SSR)
The <MUIPlacesAutocomplete> component was incorrectly creating a new
instance of the autocomplete service client in the
'componentWillMount()' lifecycle method. While this worked on the
client-side it didn't on the server-side. That's because the
'componentWillMount()' lifecycle method is the only lifecycle method
called during server-side rendering (i.e.
'ReactDOMServer.renderToString()') and when it is called the 'window'
object doesn't exist.

To fix the problem we moved the creation of the autocomplete service
client to the 'componentDidMount()' lifecycle method.
	modified:   README.md
	modified:   src/MUIPlacesAutocomplete.jsx