Prepare package/module for publishing (#5)
* Point package entry point to 'dist/index.js'

While testing if the package is ready to be published it was found that
the entry point for the package wasn't sufficient. It was pointing to
'src/index.js' which caused other projects to have compilation errors as
some of the syntax was unknown (as it was ES6 syntax). By changing the
package entry point to 'dist/index.js' which contains the transpiled
source the module should be ready for immediate usage after installing
the package.
	modified:   package.json

* Make <MUIPlacesAutocomplete> safe to render on the server (SSR)

The <MUIPlacesAutocomplete> component was incorrectly creating a new
instance of the autocomplete service client in the
'componentWillMount()' lifecycle method. While this worked on the
client-side it didn't on the server-side. That's because the
'componentWillMount()' lifecycle method is the only lifecycle method
called during server-side rendering (i.e.
'ReactDOMServer.renderToString()') and when it is called the 'window'
object doesn't exist.

To fix the problem we moved the creation of the autocomplete service
client to the 'componentDidMount()' lifecycle method.
	modified:   README.md
	modified:   src/MUIPlacesAutocomplete.jsx

* Test and remove 'dist/' folder as part of 'prepublish' script/event

The 'prepublish' script/event has been updated to run our tests one last
time before we publish. We also remove the 'dist/' folder to ensure we
don't distribute any old bits.
	modified:   package.json