Use latest Node 8 release with Travis CI build
Previously I used the latest Node 6 release with the Travis CI build as
this was what I had running locally. I realized that what we would
deploy would run on the latest Node so I upgrade locally to the latest
Node which is 8. Also 8 is the most recent Node version to be under LTS.
As a result I've now changed the Travis CI build to use Node 8 as well
to match what we will deploy under and what we ought to be using locally
for development.
	modified:   .travis.yml