Chore: Add test coverage to verify rendered suggestions are unique
This commit adds test coverage to verify that non-unique predictions
returned by the Google Maps Autocomplete Service are rendered as unique
suggestions.

We refactored the code some by moving the logic of creating a list of
unique suggestions out of the 'onInputValueChange()' method into the
'renderSuggestionsContainer()' method. I did this as it allows for more
clear testing of ideas. While adding test coverage I was conflating the
issues of having predictions returned async/React setting state async
and having all rendered suggestions be unique. This caused me to try
some weird testing patterns where I was waiting for the async React
state setting to complete and then verify my suggestions.

It seems doing this is less performant/optimized but maybe I was being
premature in that regard and this is going to lead to better
debugging/maintenance.
	modified:   src/MUIPlacesAutocomplete.jsx
	modified:   test/test.jsx
	modified:   test/test.jsx.snap