Chore: Update demo to make use of 'renderTarget' prop
Commit e40dbec5ba08bb3b8d680da4332afe0bfaa8b0e7 added a new library
named 'react-popper' (wrapper around PopperJS) to help us render
over/popover the list of suggestions over a component/element. It also
added a prop to the <MUIPlacesAutocomplete> component named
'renderTarget' that renders said components/elements to render the list
of suggestions over.

This commit updates the demo to make use of the newly required
'renderTarget' prop. We simply provided a self-closing <div> which is
unexciting. In the future maybe we can invest sometime in the demo to
provide something like a form for providing address information that
autofills when a suggestion is selected.
	modified:   demo/Demo.jsx