Docs: Format props table in README
This commit is really meant to help us trigger a publish of our package
via 'semantic-release'. Merge #15 (commit
4ec410ad681022316977dec78875b7e6e4d9fb36) made fixes (issue #13 & #14)
which would dictate we need to rev our version according to semver
rules. The individual commits that made fixes followed proper ESLint
formatting but the MR title didn't. Since we are using squash and merge
'semantic-release' didn't think the merge should be cause for
publishing.
	modified:   README.md