New: Fire callback when suggestion selected (fixes #17)
This commit implements functionality to fire a callback when a
suggestion is selected. Consumers will now provide a callback to a new
prop named 'onSuggestionSelected' which will be fired with the selected
suggestion.

This functionality ought to have been implemented before we released
<MUIPlacesAutocomplete> but it slipped my mind. Now the component ought
to be somewhat useful for others to start using.
	modified:   .eslintrc
	modified:   README.md
	modified:   demo/Demo.jsx
	modified:   package.json
	modified:   src/MUIPlacesAutocomplete.jsx
	modified:   test/test.jsx