New: Allow the <input> element to be controlled (fixes #21) (#23)
* New: Allow the <input> element to be controlled (fixes #21)

This commit implements functionality that allows the <input> element
rendered by the MUI <TextField> component to be controlled externally.
This is done by passing the value of the <input> element as the
'inputValue' prop on the <Downshift> component we render.

This commit satisfies the work of: #21
	modified:   README.md
	modified:   demo/Demo.jsx
	copied:     demo/Demo.jsx -> demo/DemoBasic.jsx
	new file:   demo/DemoControlledInput.jsx
	new file:   demo/rootReducer.js
	modified:   package.json
	modified:   src/MUIPlacesAutocomplete.jsx
	modified:   test/test.jsx
	modified:   yarn.lock

* New: Allow the <input> element to be controlled, part deux (fixes #21)

PR #22 sought to introduce new functionality that would allow the
<input> element rendered by the MUI <TextField> component to be
controlled externally. After some initial feedback and talking to Matt
about it some it was obvious that I went about implementing this
functionality incorrectly in commit
e27d46596523f0206fef05b00641fe94f7c1c39e. After re-reading the docs for
Downshift and MUI <TextField> this commit re-implements the
functionality in a more consistent fashion with how Downshift/MUI are
documented/expected to be used.
	modified:   README.md
	modified:   demo/DemoControlledInput.jsx
	modified:   src/MUIPlacesAutocomplete.jsx
	modified:   test/test.jsx