Fix: Spread 'textFieldProps' last for 'getInputProps()' (fixes #24) (#25)
Bug #24 discussed that the 'id' for the rendered <input> element
couldn't be set contrary to what documentation said. The propblem was we
were overwriting the user input for the 'id' value by spreading
'textFieldProps' onto the object passed to 'getInputProps()' first and
then applying our own 'id'. This commit fixes the issue by spreading the
'textFieldProps' lastly after we set our own 'id'.
	modified:   src/MUIPlacesAutocomplete.jsx
	modified:   test/test.jsx