New: Expose 'createAutocompleteRequest' prop (fixes #26)
This commit exposes a new prop on <MUIPlacesAutocomplete> named
'createAutocompleteRequest'. Consumers of <MUIPlacesAutocomplete> may
provide a function to the 'createAutocompleteRequest' prop that returns
an object that is used as the request to the Google Places API Web
Service when querying for suggestions. This object may be customized
(per Google API docs) to allow for greater control over what suggestions
are returned to us (the user).
	modified:   README.md
	modified:   src/MUIPlacesAutocomplete.jsx