Fix: Set 'z-index' on Popper to render over descendants (fixes #28)
Suggestions weren't being rendered over descendant elements/components
to <MUIPlacesAutocomplete>. This was due to a lack of understanding
about how stacking worked. After learning more about it on MDN we set
the 'z-index' on the <Popper> which provides the modality for our
rendered suggestions. This fixes the problem as descendant
elements/components will be in the same stacking context as <MPA>.

We provide a default 'z-index' of 1. Looking forward we ought to allow
people to customize the 'z-index' further. Rather than solving that
issue in this commit we will do so in another that allows for fine
granular control over <Popper> as well as the other components that
<MPA> composes.
	modified:   src/MUIPlacesAutocomplete.jsx
	modified:   test/test.jsx