Upgrade: Rev dependencies (fixes #31) (#32)
* Upgrade: Rev 'downshift' from 1.22.5 to 1.31.2

This commit revs the 'downshift' package version form 1.22.5 to 1.31.2.
After reading the release docs it didn't appear that there was any
breaking changes. Tests passed and functionality appears okay.
	modified:   test/api-key.js
	modified:   test/test.jsx.snap
	modified:   yarn.lock

* Upgrade: Rev 'react-popper' from 0.7.4 to 0.8.3

This commit revs the 'react-popper' package version from 0.7.4 to 0.8.3.
After reading the release docs it didn't appear that there were any
breaking changes. Tests passed and functionality appears okay.
	modified:   package.json
	modified:   yarn.lock

* Upgrade: Rev 'autosuggest-highlight' from 3.1.0 to 3.1.1

This commit revs the 'autosuggest-highlight' package version from 3.1.0
to 3.1.1. After reading the release docs it didn't appear that there
were any breaking changes. Tests passed and functionality appears okay.
	modified:   yarn.lock

* Upgrade: Rev 'material-ui' from 1.0.0-beta.22 to 1.0.0-beta.38

This commit revs the 'material-ui' package version from 1.0.0-beta.22 to
1.0.0-beta.38. After reading the release docs there were some breaking
changes we had to account for but only in the demo code.

The <Snackbar> we use in the demo code needs to now use 'onClose'
instead of 'onRequestClose'. This is a result of the 1.0.0-beta.24
release of 'material-ui'. Here are the release notes/PR that describe
the changes:
* https://github.com/mui-org/material-ui/releases/tag/v1.0.0-beta.24
* https://github.com/mui-org/material-ui/pull/9451

The <Button> we use in the dmeo code needs to now use the 'variant' prop
instead of the 'raised' prop to describe what type of <Button> it is.
This is a result of the 1.0.0-beta.32 release of 'material-ui'. Here are
the release notes/API guide that explains the changes:
* https://github.com/mui-org/material-ui/releases/tag/v1.0.0-beta.32
* https://material-ui-next.com/guides/api/#boolean-vs-enum

After accounting for the breaking changes functionality in both the demo
code and <MUIPlacesAutocomplete> code appears okay. Tests are also
passing.
	modified:   demo/DemoBasic.jsx
	modified:   demo/DemoControlledInput.jsx
	modified:   test/test.jsx.snap
	modified:   yarn.lock

* Docs: Add a "preview" .gif of component in action to README.md
	modified:   README.md

* Docs: Fix Travis CI image/badge link
	modified:   README.md