Breaking: Upgrade to MUI v1 stable release (fixes #38) (#39)
This commit upgrades to the first stable release of MUI (yay!). This
commit accounts for breaking changes made in v1.0.0 of MUI. We mainly
had to account for how MUI components were imported now as well as
upgrade to React 16.3 as it is a requirement for MUI v1.0.0 now. As a
result we are also declaring this commit to be a breaking change. This
is to account for the fact that the peer dependencies have changed for
MUI as well as what versions of the peer dependencies are required for
MUI and React.

Thanks to @charliedotau and @charlax on the heads up about the new
version of MUI and the breaking changes we need to account for.